table css

+++ CourseWork/UltraPioner/Models/DataBase/Entities/Standart.cs
@@ -4,9 +4,8 @@
 	{
 		public int Id { get; set; }
 		public string StandartName { get; set; }
-		public int StandartResult { get; set; }
+		public double StandartResult { get; set; }
 		public DateTime DateResult { get; set; }
-
 		//
 		//public int TrainerMagazineID { get; set; }
 		//public TrainerMagazine TrainerJurnal { get; set; }
